backend/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
flask-session
redis
python-dotenv
mysqlclient
flask-marshmallow
marshmallow-sqlalchemy
flask-cors
flask-mail
pysha3
requests
psycopg2
gunicorn
flask-mysql
mysql-connector-python
